Remove pirated ArrowTypes/TimeSpans interop
Onda currently defines Arrow (de)serialization for `TimeSpan`. This
functionality is being moved to an ArrowTypes extension for TimeSpans,
which will be released in TimeSpans 1.1.0. Having the same methods
defined in Onda and the extension causes precompilation to fail. To fix
this, we can simply delete the pirated interoperability between
ArrowTypes and TimeSpans and set the TimeSpans compatibility to 1.1.0.