Allow `convert`ing encoded samples if the element type matches (#171)
At present, we error on `convert(::Type{Samples{A}}, ::Samples{B})` for
all `A` and `B` if the samples are encoded. This is reasonable since we
don't want to affect how the samples get decoded. However, decoding
should be based on the type of the encoded samples themselves rather
than of their container; a decoder shouldn't care whether the incoming
data is e.g. a full matrix or a view of a matrix. Thus we can allow
conversion for encoded samples provided that only the container type is
affected, not the element type.